StakeTxDetails component test
